+++ editor/src/components/filebrowser/confirm-close-dialog.tsx
@@ -5,7 +5,7 @@ import { FormButton } from 'uuiui'
 import { Dialog } from 'uuiui'
 import { EditorDispatch } from '../editor/action-types'
 import * as EditorActions from '../editor/actions/actions'
-import { EditorTab, isOpenFileTab } from '../editor/store/editor-state'
+import { EditorTab, isOpenFileTab } from '../editor/store/editor-tabs'
 
 interface ConfirmCloseDialogProps {
   dispatch: EditorDispatch
