+++ editor/src/components/filebrowser/fileitem.tsx
@@ -23,7 +23,7 @@ import { ContextMenuItem, requireDispatch } from '../context-menu-items'
 import { ContextMenuWrapper } from '../context-menu-wrapper'
 import { EditorAction, EditorDispatch } from '../editor/action-types'
 import * as EditorActions from '../editor/actions/actions'
-import { openFileTab } from '../editor/store/editor-state'
+import { openFileTab } from '../editor/store/editor-tabs'
 import { ExpandableIndicator } from '../navigator/navigator-item/expandable-indicator'
 import { FileBrowserItemInfo, FileBrowserItemType } from './filebrowser'
 import { PasteResult } from '../../utils/clipboard-utils'
