+++ editor/src/components/inspector/sections/component-section/component-section.tsx
@@ -32,6 +32,7 @@ import { foldEither } from '../../../../core/shared/either'
 import { mapToArray } from '../../../../core/shared/object-utils'
 import { PropertyPath } from '../../../../core/shared/project-file-types'
 import * as PP from '../../../../core/shared/property-path'
+import { useKeepReferenceEqualityIfPossible } from '../../../../utils/react-performance'
 import Utils from '../../../../utils/utils'
 import { getParseErrorDetails, ParseError } from '../../../../utils/value-parser-utils'
 import { InfoBox } from '../../../common/notices'
@@ -45,7 +46,6 @@ import {
   useInspectorInfoForPropertyControl,
 } from '../../common/property-controls-hooks'
 import {
-  useKeepReferenceEqualityIfPossible,
   useSelectedPropertyControls,
   useUsedPropsWithoutControls,
   useUsedPropsWithoutDefaults,
